requirements.txt
psutil
os
json
speedtest-cli